requirements/basic.txt
click>=6.7
flatten-dict>=0.0.2
graphviz>=0.8.3
numpy>=1.14.3
pandas>=0.22.0
Pillow>=5.1.0
pygraphviz>=1.3.1
scikit-learn>=0.19.1
scipy>=1.0.1
tqdm>=4.23.2
torch>=0.4.0
tensorboardX>=1.2
PyYAML>=3.12
